ENH: move SpecificationError from core/base.py to error/__init__.py per GH27656
