+++ doc/source/reference/testing.rst
@@ -41,6 +41,7 @@ Exceptions and warnings
    errors.ParserError
    errors.ParserWarning
    errors.PerformanceWarning
+   errors.SpecificationError
    errors.UnsortedIndexError
    errors.UnsupportedFunctionCall
