+++ pandas/errors/__init__.py
@@ -244,3 +244,27 @@ class DataError(Exception):
     Or, it can be raised when trying to apply a function to a non-numerical
     column on a rolling window.
     """
+
+
+class SpecificationError(Exception):
+    """
+    Exception raised in two scenarios. The first way is calling agg on a
+    Dataframe or Series using a nested renamer (dict-of-dict).
+    The second way is calling agg on a Dataframe with duplicated functions
+    names without assigning column name.
+
+    Examples
+    --------
+    >>> df = pd.DataFrame({'A': [1, 1, 1, 2, 2],
+    ...                    'B': range(5),
+    ...                    'C':range(5)})
+    >>> df.groupby('A').B.agg({'foo': 'count'})
+    ... # SpecificationError
+
+    >>> df.groupby('A').agg({'B': {'foo': ['sum', 'max']},
+    ...                      'C': {'bar': ['count', 'min']}})
+    ... # SpecificationError
+
+    >>> df.groupby('A').agg(['min', 'min'])
+    ... # SpecificationError
+    """
