+++ pandas/tests/groupby/aggregate/test_aggregate.py
@@ -9,6 +9,8 @@ import re
 import numpy as np
 import pytest
 
+from pandas.errors import SpecificationError
+
 from pandas.core.dtypes.common import is_integer_dtype
 
 import pandas as pd
@@ -21,7 +23,6 @@ from pandas import (
     to_datetime,
 )
 import pandas._testing as tm
-from pandas.core.base import SpecificationError
 from pandas.core.groupby.grouper import Grouping
