+++ pandas/tests/groupby/aggregate/test_other.py
@@ -8,6 +8,8 @@ from functools import partial
 import numpy as np
 import pytest
 
+from pandas.errors import SpecificationError
+
 import pandas as pd
 from pandas import (
     DataFrame,
@@ -19,7 +21,6 @@ from pandas import (
     period_range,
 )
 import pandas._testing as tm
-from pandas.core.base import SpecificationError
 
 from pandas.io.formats.printing import pprint_thing
